Move command to a separated file from the cli
